requirements.txt
aiohttp==3.7.4.post0
asciichartpy==1.5.25
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
dateparser==1.0.0
Flask==2.0.0
halo==0.0.31
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
log-symbols==0.0.14
MarkupSafe==2.0.0
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
spinners==0.0.24
termcolor==1.1.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websockets==9.0.1
Werkzeug==2.0.0
yarl==1.6.3
